+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Making placeholder image of video block to take 100% width when it is right or left aligned @iFlameing
 - Showing clear icon when title is too long in objectbrowser selected items in multiple mode @iFlameing
 - Set default value selected for variation in listing block @giuliaghisini
+- Use querystring prop in ListingBody @giuliaghisini
 
 ### Internal